refactor(test): move format tests to new file
